Move delaying code to enqueueing service
