+++ services/enqueueing/enqueueing.go
@@ -1,4 +1,4 @@
-package queueing
+package enqueueing
 
 import (
 	"errors"
@@ -17,7 +17,7 @@ func init() {
 	enqueuer = work.NewEnqueuer("google_scraper", database.GetRedisPool())
 }
 
-func AddToQueue(keyword *models.Keyword, secondsInTheFuture int64) (*work.ScheduledJob, error) {
+func DelayedEnqueue(keyword *models.Keyword, keywordIndex int64) (*work.ScheduledJob, error) {
 	if keyword.Id <= 0 {
 		return nil, errors.New("invalid keyword object")
 	}
@@ -28,11 +28,18 @@ func AddToQueue(keyword *models.Keyword, secondsInTheFuture int64) (*work.Schedu
 		return nil, err
 	}
 
-	job, err := enqueuer.EnqueueIn(crawlingJobName, secondsInTheFuture, work.Q{"keywordId": keyword.Id})
+	delayTimeInSeconds := getDelayTimeInSeconds(keywordIndex)
+	job, err := enqueuer.EnqueueIn(crawlingJobName, delayTimeInSeconds, work.Q{"keywordId": keyword.Id})
 
 	if err != nil {
 		return nil, err
 	}
 
 	return job, nil
+}
+
+func getDelayTimeInSeconds(keywordIndex int64) int64 {
+	// Each job will be run two seconds later than the previous job, jobs will be enqueued immediately
+	// But will be run based on this seconds value in the future
+	return keywordIndex + 2
 }
